+++ src/util/performance/performance_data.h
@@ -8,14 +8,13 @@ template <class T>
 class PerformanceData {
 private:
     std::deque<T> durations_;
-    int period_;
+    size_t period_;
     T total_;
 
 public:
-    PerformanceData(int period = 60): period_(period) {
+    PerformanceData(size_t period = 60): period_(period) {
         total_ = T(0);
     }
-
     void Add(T duration) {
         total_ += duration;
         durations_.push_back(duration);
